+++ packages/cli/src/command/init/generateNpmName.ts
@@ -1,0 +1,7 @@
+import * as decamelize from 'decamelize';
+
+export default function (name: string, npmScope?: string): string {
+  // WebkitTransform -> webkit-transform
+  name = decamelize(name, '-');
+  return npmScope ? `${npmScope}/${name}` : name;
+}