+++ packages/doctor/src/config.ts
@@ -3,6 +3,7 @@ export default {
     files: 'files.json',
     report: {
       eslint: 'eslint-result.json',
+      stylelint: 'stylelint-result.json',
       escomplex: 'escomplex-result.json',
       jscpd: 'jscpd-result.json',
       codemod: 'codemod-result.json',