+++ packages/doctor/src/types/Scanner.ts
@@ -1,44 +1,24 @@
 import { IClone } from '@jscpd/core';
-import { IResult } from '@appworks/codemod';
+import type { ProjectLintResult, CodemodRule } from '@applint/applint';
 
 export interface IScannerOptions {
-  ignore: string[];
+  ignore?: string[];
 }
 
 export interface IScanOptions {
   fix?: boolean;
   framework?: string;
-  transforms?: string[];
+  transforms?: Record<string, number>;
   languageType?: 'js' | 'ts';
   tempFileDir?: string;
   timeout?: number;
   disableESLint?: boolean;
+  disableStylelint?: boolean;
   disableMaintainability?: boolean;
   disableRepeatability?: boolean;
   disableCodemod?: boolean;
   maxRepeatabilityCheckLines?: number;
-}
-
-// https://www.npmjs.com/package/typhonjs-escomplex
-export interface IMaintainabilityReport {
-  classes: any[];
-  errors: any[];
-  methods: any[];
-  aggregate: any;
-  aggregateAverage: any;
-  methodAverage: any;
-  settings: any;
-  srcPathAlias: any;
-  filePath: string;
-  srcPath: string;
-  lineEnd: number;
-  lineStart: number;
-  maintainability: number;
-}
-
-export interface IMaintainabilityReports {
-  score: number;
-  reports: IMaintainabilityReport[];
+  customTransformRules?: Record<string, CodemodRule>;
 }
 
 export interface IRepeatabilityReports {
@@ -46,7 +26,7 @@ export interface IRepeatabilityReports {
   clones: IClone[];
 }
 
-export interface IEslintReports {
+export interface IESLintReports {
   score: number;
   reports: any[];
   errorCount: number;
@@ -54,9 +34,16 @@ export interface IEslintReports {
   customConfig: any;
 }
 
+export interface IStylelintReports {
+  score: number;
+  reports: any[];
+  warningCount: number;
+  customConfig: any;
+}
+
 export interface ICodemodReports {
   score: number;
-  reports: IResult[];
+  reports: ProjectLintResult;
 }
 
 export interface IScannerReports {
@@ -66,8 +53,10 @@ export interface IScannerReports {
   };
   score?: number;
   scanTime?: number;
-  ESLint?: IEslintReports;
-  maintainability?: IMaintainabilityReports;
+  ESLint?: IESLintReports;
+  Stylelint?: IStylelintReports;
   repeatability?: IRepeatabilityReports;
   codemod?: ICodemodReports;
+  // the content is empty
+  maintainability?: any;
 }