+++ packages/doctor/src/types/index.ts
@@ -1,0 +1,5 @@
+export * from './Analyzer';
+export * from './Doctor';
+export * from './File';
+export * from './Scanner';
+export * from './Scorer';