+++ packages/doctor/CHANGELOG.md
@@ -1,5 +1,17 @@
 # changelog
 
+## 0.4.0
+
+- feat: support scan and fix css files
+
+## 0.3.1
+
+- chore: remove maintainability detail from report and `typhonjs-escomplex` dependency
+
+## 0.3.0
+
+- Feat: replace `@appworks/codemod` with `@applint/applint`
+
 ## 0.2.3
 
 - Fix Windows get empty files